pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fullstackautoquant"
version = "0.1.0"
description = "End-to-End Deep Learning Quantitative Trading System"
readme = "README.md"
license = {text = "CC-BY-NC-SA-4.0"}
requires-python = ">=3.10"
authors = [
    {name = "Zizhao Huang", email = "zizhaohuang26@gmail.com"},
]
keywords = [
    "quantitative-trading",
    "deep-learning",
    "stock-ranking",
    "automated-trading",
    "risk-management",
    "backtesting",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pandas>=1.5",
    "pyqlib>=0.9.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "h5py>=3.0",
    "python-dotenv>=1.0",
    "tushare>=1.3",
]

[project.optional-dependencies]
trading = [
    "gmtrade",
]
webui = [
    "streamlit>=1.20",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "black>=23.0",
    "isort>=5.0",
    "pre-commit>=3.0",
]
all = [
    "fullstackautoquant[trading,webui,dev]",
]

[project.urls]
Homepage = "https://github.com/Zizhao-HUANG/FullStackAutoQuant"
Repository = "https://github.com/Zizhao-HUANG/FullStackAutoQuant"
Issues = "https://github.com/Zizhao-HUANG/FullStackAutoQuant/issues"

# ======================== Tool Configuration ========================

[tool.setuptools.packages.find]
include = ["fullstackautoquant*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["fullstackautoquant"]
skip = [".git", "__pycache__", "tmp", "logs"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["fullstackautoquant", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # bugbear
    "SIM",  # simplify
]
ignore = [
    "E203",
    "E266",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
